Fix setting the AnnotWidget of standalone fields
When loading standalone Formfields, we were wrongly
creating new AnnotWidgets to fill its FormWidget member,
whereas we just needed to set it to the own AnnotWidget
that contains the standalone field.

That meant that when clicking on standalone checkbox
fields we were really clicking on AnnotWidgets that
were not the ones used for drawing the page.

With this fix those checkboxes now work ok.

Fixes #642
Improves #1055